+++ rally_openstack/services/storage/cinder_v3.py
@@ -33,7 +33,7 @@ class CinderV3Service(service.Service, cinder_common.CinderMixin):
                       snapshot_id=None, source_volid=None, name=None,
                       description=None, volume_type=None,
                       availability_zone=None, metadata=None, imageRef=None,
-                      scheduler_hints=None, multiattach=False, backup_id=None):
+                      scheduler_hints=None, backup_id=None):
         """Creates a volume.
 
         :param size: Size of volume in GB
@@ -48,8 +48,6 @@ class CinderV3Service(service.Service, cinder_common.CinderMixin):
         :param source_volid: ID of source volume to clone from
         :param scheduler_hints: (optional extension) arbitrary key-value pairs
                             specified by the client to help boot an instance
-        :param multiattach: Allow the volume to be attached to more than
-                            one instance
         :param backup_id: ID of the backup
 
         :returns: Return a new volume.
@@ -64,7 +62,6 @@ class CinderV3Service(service.Service, cinder_common.CinderMixin):
                   "metadata": metadata,
                   "imageRef": imageRef,
                   "scheduler_hints": scheduler_hints,
-                  "multiattach": multiattach,
                   "backup_id": backup_id}
         if isinstance(size, dict):
             size = random.randint(size["min"], size["max"])
@@ -100,11 +97,11 @@ class CinderV3Service(service.Service, cinder_common.CinderMixin):
 
     @atomic.action_timer("cinder_v3.list_volumes")
     def list_volumes(self, detailed=True, search_opts=None, marker=None,
-                     limit=None, sort_key=None, sort_dir=None, sort=None):
+                     limit=None, sort=None):
         """List all volumes."""
         return self._get_client().volumes.list(
             detailed=detailed, search_opts=search_opts, marker=marker,
-            limit=limit, sort_key=sort_key, sort_dir=sort_dir, sort=sort)
+            limit=limit, sort=sort)
 
     @atomic.action_timer("cinder_v3.list_types")
     def list_types(self, search_opts=None, is_public=None):
@@ -246,7 +243,7 @@ class UnifiedCinderV3Service(cinder_common.UnifiedCinderMixin,
                       volume_type=None, user_id=None,
                       project_id=None, availability_zone=None,
                       metadata=None, imageRef=None, scheduler_hints=None,
-                      source_replica=None, multiattach=False, backup_id=None):
+                      source_replica=None, backup_id=None):
         """Creates a volume.
 
         :param size: Size of volume in GB
@@ -264,8 +261,6 @@ class UnifiedCinderV3Service(cinder_common.UnifiedCinderMixin,
         :param source_volid: ID of source volume to clone from
         :param scheduler_hints: (optional extension) arbitrary key-value pairs
                             specified by the client to help boot an instance
-        :param multiattach: Allow the volume to be attached to more than
-                            one instance
         :param backup_id: ID of the backup
 
         :returns: Return a new volume.
@@ -277,10 +272,10 @@ class UnifiedCinderV3Service(cinder_common.UnifiedCinderMixin,
             description=description, volume_type=volume_type,
             availability_zone=availability_zone, metadata=metadata,
             imageRef=imageRef, scheduler_hints=scheduler_hints,
-            multiattach=multiattach, backup_id=backup_id))
+            backup_id=backup_id))
 
     def list_volumes(self, detailed=True, search_opts=None, marker=None,
-                     limit=None, sort_key=None, sort_dir=None, sort=None):
+                     limit=None, sort=None):
         """Lists all volumes.
 
         :param detailed: Whether to return detailed volume info.
@@ -288,16 +283,13 @@ class UnifiedCinderV3Service(cinder_common.UnifiedCinderMixin,
         :param marker: Begin returning volumes that appear later in the volume
                        list than that represented by this volume id.
         :param limit: Maximum number of volumes to return.
-        :param sort_key: Key to be sorted; deprecated in kilo
-        :param sort_dir: Sort direction, should be 'desc' or 'asc'; deprecated
-                         in kilo
         :param sort: Sort information
         :returns: Return volumes list.
         """
         return [self._unify_volume(volume)
                 for volume in self._impl.list_volumes(
                 detailed=detailed, search_opts=search_opts, marker=marker,
-                limit=limit, sort_key=sort_key, sort_dir=sort_dir, sort=sort)]
+                limit=limit, sort=sort)]
 
     def get_volume(self, volume_id):
         """Get a volume.