+++ tests/unit/contexts/dataplane/test_heat.py
@@ -17,7 +17,7 @@ import functools
 
 import mock
 
-from rally_openstack.contexts import dataplane
+from rally_openstack.contexts.dataplane import heat as heat_dataplane
 from tests.unit import test
 
 MOD = "rally_openstack.contexts.dataplane.heat."
@@ -28,13 +28,13 @@ class TestHeatWorkload(test.ScenarioTestCase):
     @mock.patch(MOD + "pkgutil")
     def test_get_data_resource(self, mock_pkgutil):
         mock_pkgutil.get_data.return_value = "fake_data"
-        data = dataplane.heat.get_data([1, 2])
+        data = heat_dataplane.get_data([1, 2])
         self.assertEqual("fake_data", data)
         mock_pkgutil.get_data.assert_called_once_with(1, 2)
 
     @mock.patch(MOD + "open")
     def test_get_data_file(self, mock_open):
-        data = dataplane.heat.get_data(1)
+        data = heat_dataplane.get_data(1)
         self.assertEqual(mock_open.return_value.read.return_value, data)
         mock_open.assert_called_once_with(1)
 
@@ -42,7 +42,7 @@ class TestHeatWorkload(test.ScenarioTestCase):
         user = [1, 2]
         tenant = [3, 4, {"one": 1}]
         self.context["tenants"] = {1: tenant}
-        ctx = dataplane.heat.HeatDataplane(self.context)
+        ctx = heat_dataplane.HeatDataplane(self.context)
         gcp = functools.partial(ctx._get_context_parameter, user, 1)
         self.assertEqual(1, gcp("user.0"))
         self.assertEqual(2, gcp("user.1"))
@@ -58,7 +58,7 @@ class TestHeatWorkload(test.ScenarioTestCase):
         mock_clients.return_value = mock.Mock(
             neutron=mock.Mock(return_value=fake_nc))
         self.context["admin"] = {"credential": "fake_credential"}
-        ctx = dataplane.heat.HeatDataplane(self.context)
+        ctx = heat_dataplane.HeatDataplane(self.context)
         network_id = ctx._get_public_network_id()
         self.assertEqual("fake_id", network_id)
         mock_clients.assert_called_once_with("fake_credential")
@@ -85,7 +85,7 @@ class TestHeatWorkload(test.ScenarioTestCase):
         })
         mock_heat_dataplane__get_context_parameter.return_value = "gcp"
         mock_get_data.side_effect = ["tpl", "sf1", "sf2"]
-        ctx = dataplane.heat.HeatDataplane(self.context)
+        ctx = heat_dataplane.HeatDataplane(self.context)
         ctx._get_public_network_id = mock.Mock(return_value="fake_net")
         ctx.setup()
         workloads = self.context["tenants"]["t1"]["stack_dataplane"]